Store default start day in a class level instance variable
Previously, a class variable was used to set the start day for
OrderedWeek. Such variables have a tendency to produce unforeseen
effects, especially with inheritance. Replace class variable with class
level instance variable. Add inherited hook to OrderedWeek, which sets
the default start day for a subclass when it inherits from OrderedWeek.
This allows subclasses to manage their own start day, separate from that
of the super class.