Raise exceptions for invalid arguments
Instead of silently failing when updating the start day, and defaulting
to this week if an invalid date is given, raise exceptions to telling
the user when something unexpected is passed. Add OrderedWeek::WeekDay
inner class to hold the list of day names and handle validating
start_day configurations (automatically converts strings to symbols and
handles capitalization).